requirements.txt
requests==2.32.3
environs==11.2.1
python-telegram-bot==13.15
urllib3==1.26.15